requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.6.2
distro==1.9.0
Django==5.0.6
django-environ==0.11.2
EbookLib==0.18
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
lxml==5.2.2
markdown-it-py==3.0.0
mdurl==0.1.2
mysqlclient==2.2.4
openai==1.34.0
packaging==24.1
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
rich==13.7.1
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.2
wheel==0.43.0
